Split HabitRepository into HabitRepository and EntryRepository
