+++ app/src/main/java/com/willbsp/habits/data/repo/HabitRepository.kt
@@ -1,6 +1,5 @@
 package com.willbsp.habits.data.repo
 
-import com.willbsp.habits.data.model.Entry
 import com.willbsp.habits.data.model.Habit
 import com.willbsp.habits.data.model.HabitWithEntries
 import kotlinx.coroutines.flow.Flow
@@ -12,13 +11,9 @@ interface HabitRepository {
 
     suspend fun getHabitById(habitId: Int): Habit
 
-    suspend fun getEntryForDate(date: LocalDate, habitId: Int): Entry?
-
     suspend fun addHabit(habit: Habit)
 
     suspend fun updateHabit(habit: Habit)
-
-    suspend fun toggleEntry(habitId: Int, date: LocalDate)
 
     suspend fun deleteHabit(habitId: Int)
