+++ app/src/main/java/com/willbsp/habits/data/repo/LocalHabitRepository.kt
@@ -1,8 +1,6 @@
 package com.willbsp.habits.data.repo
 
-import com.willbsp.habits.data.database.dao.EntryDao
 import com.willbsp.habits.data.database.dao.HabitDao
-import com.willbsp.habits.data.model.Entry
 import com.willbsp.habits.data.model.Habit
 import com.willbsp.habits.data.model.HabitWithEntries
 import kotlinx.coroutines.flow.Flow
@@ -10,8 +8,7 @@ import java.time.LocalDate
 import javax.inject.Inject
 
 class LocalHabitRepository @Inject constructor(
-    private val habitDao: HabitDao,
-    private val entryDao: EntryDao
+    private val habitDao: HabitDao
 ) : HabitRepository {
 
     override fun getAllHabitsWithEntriesForDates(dates: List<LocalDate>): Flow<List<HabitWithEntries>> {
@@ -23,33 +20,12 @@ class LocalHabitRepository @Inject constructor(
         return Habit(habit.id, habit.name, habit.frequency)
     }
 
-    override suspend fun getEntryForDate(date: LocalDate, habitId: Int): Entry? {
-        return entryDao.getEntryForDate(date, habitId)
-    }
-
     override suspend fun addHabit(habit: Habit) {
         habitDao.insert(Habit(habit.id, habit.name, habit.frequency))
     }
 
     override suspend fun updateHabit(habit: Habit) {
         habitDao.update(Habit(habit.id, habit.name, habit.frequency))
-    }
-
-    override suspend fun toggleEntry(
-        habitId: Int,
-        date: LocalDate
-    ) { // TODO check if habit exists first!
-
-        val entry: Entry? =
-            entryDao.getEntryForDate(date, habitId)
-        if (entry == null) entryDao.insert(
-            Entry(
-                habitId = habitId,
-                date = date
-            )
-        )
-        else entryDao.delete(entry)
-
     }
 
     override suspend fun deleteHabit(habitId: Int) {
