+++ app/src/main/java/com/willbsp/habits/di/RepositoryModule.kt
@@ -1,6 +1,8 @@
 package com.willbsp.habits.di
 
+import com.willbsp.habits.data.repo.EntryRepository
 import com.willbsp.habits.data.repo.HabitRepository
+import com.willbsp.habits.data.repo.LocalEntryRepository
 import com.willbsp.habits.data.repo.LocalHabitRepository
 import dagger.Binds
 import dagger.Module
@@ -10,12 +12,18 @@ import dagger.hilt.android.scopes.ViewModelScoped
 
 @Module
 @InstallIn(ViewModelComponent::class)
-abstract class HabitsRepositoryModule {
+abstract class RepositoryModule {
 
     @Binds
     @ViewModelScoped
     abstract fun bindHabitRepository(
         localHabitRepository: LocalHabitRepository
     ): HabitRepository
+
+    @Binds
+    @ViewModelScoped
+    abstract fun bindEntryRepository(
+        localEntryRepository: LocalEntryRepository
+    ): EntryRepository
 
 }
